# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Canadian Solar ERP API",
    "description": "API complète pour la gestion de magasin solaire et électronique",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "subtotal": {
            "type": "number",
            "format": "decimal"
          },
          "totalAmount": {
            "type": "number",
            "format": "decimal"
          },
          "paidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "partially_paid",
              "cancelled",
              "shipped",
              "delivered"
            ]
          },
          "paymentMethod": {
            "type": "string"
          },
          "deliveryAddress": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Lister les commandes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commandes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Créer une commande",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commande créée"
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Récupérer une commande",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commande trouvée"
          },
          "404": {
            "description": "Commande non trouvée"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Mettre à jour une commande",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commande mise à jour"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Annuler une commande et restaurer le stock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commande annulée, stock restauré"
          },
          "404": {
            "description": "Commande non trouvée"
          }
        }
      }
    },
...
    "/api/customers/{id}/orders": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Commandes d'un client",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commandes du client"
          }
        }
      }
    },
...
    "/api/orders/customer/{customerId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Commandes d'un client",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commandes du client"
          }
        }
      }
    },
...
    "/api/payments/order/{orderId}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Paiements d'une commande",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        }
      }
    },
...
  }
}